+++ tsup.config.ts
@@ -1,7 +1,7 @@
 import { defineConfig } from 'tsup'
 
 export default defineConfig({
-  entry: ['src/cli.ts'],
+  entry: ['src/main.ts'],
   format: ['esm'],
   target: 'node20',
   outDir: 'dist',
@@ -9,7 +9,4 @@ export default defineConfig({
   splitting: false,
   sourcemap: true,
   dts: false,
-  banner: {
-    js: '#!/usr/bin/env node',
-  },
 })
